Add README.md

+++ src/Transfer.java
@@ -1,8 +1,8 @@
 public class Transfer {
     public static void main(String[] args) {
-        int balance = 2_000_000_000;
+        long balance = 2_000_000_000;
         int transfer = 500_000_000;
-        int total = balance + transfer;
+        long total = balance + transfer;
         System.out.println(total);
     }
 }
